irrigation_service/pyproject.toml
[project]
name = "irrigation-service"
version = "0.1.0"
description = ""
authors = [
    {name = "Ogo",email = "ogochukwubah@gmail.com"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "paho-mqtt (>=2.1.0,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "flask-migrate (>=4.1.0,<5.0.0)",
    "flask-bcrypt (>=1.0.1,<2.0.0)",
    "flask-jwt-extended (>=4.7.1,<5.0.0)",
    "flask-restx (>=1.3.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
